Don't trigger TRAMP operations when entering minibuffer
ESS has a hook that can trigger a TRAMP operation when entering the
minibuffer for any reason. This is problematic if in a buffer for a
TRAMP file whose server is currently not accessible, since it will
cause Emacs to hang on any attempt to use the minibuffer. We add some
advice to disable this hook in the minibuffer.